AgenticWorkers/KB/requirements.txt
# Core dependencies
pymupdf>=1.24.0
sentence-transformers>=3.0.0
pinecone>=5.0.0
groq>=0.9.0
tqdm>=4.66.0
python-dotenv>=1.0.0
beautifulsoup4>=4.13.4

# Azure
azure-storage-queue>=12.9.0
azure-storage-blob>=12.19.0

# HTTP
requests>=2.32.0
httpx>=0.27.0

# ML/AI
transformers>=4.40.0
torch>=2.0.0
